+++ scripts/release-artifact-policy.cjs
@@ -1,0 +1,5 @@
+function isHtmlDocument(file) {
+  return /\.html?$/i.test(file);
+}
+
+module.exports = { isHtmlDocument };