+++ src/app/+html.tsx
@@ -11,7 +11,6 @@ export default function RootHtml({ children }: PropsWithChildren) {
           name="viewport"
           content="width=device-width, initial-scale=1, shrink-to-fit=no"
         />
-        <meta name="robots" content="noindex,nofollow,noarchive" />
         <ScrollViewStyleReset />
       </head>
       <body>{children}</body>